+++ src/kernel/kernel.c
@@ -4,8 +4,6 @@
 // kernel.c for MRNIU/SimpleKernel.
 
 #include "kernel.h"
-#include "assert.h"
-
 
 // 内核入口
 void kernel_main(uint32_t magic, uint32_t addr) {
@@ -15,6 +13,7 @@ void kernel_main(uint32_t magic, uint32_t addr) {
 		idt_init();       // IDT 初始化
 		clock_init();       // 时钟初始化
 		keyboard_init();       // 键盘初始化
+		pmm_init();
 		showinfo();
 
 		while(1);
